requirements.txt
streamlit>=1.28.0
openai>=1.50.0
yt-dlp>=2024.12.0
python-dotenv>=1.0.0